mb/biostar: Replace verb tables with reworked implementation
No modifications/assumptions about the codecs were made on these boards.

It should be noted that we do not modifiy the verb tables in any case,
as it would break the regression test script mentioned in the TEST
section below.

For an overall rationale for this rework, see commit 31fc5b06a6be
("device: Introduce reworked azalia verb table").

TEST= All boards passed regression test (CB:88763)

Change-Id: I5d814917174a02c490d9f38141b0d07f9059a8f5
Signed-off-by: Nicholas Sudsgaard <devel+coreboot@nsudsgaard.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/88663
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Felix Singer <service+coreboot-gerrit@felixsinger.de>